Reduce code duplication in tool calling examples
