+++ .scalafmt.conf
@@ -34,3 +34,29 @@ docstrings.wrap = no
 
 # Trailing commas
 trailingCommas = preserve
+fileOverride {
+  "glob:**/samples/src/main/scala/**" {
+     runner.dialect = scala3
+  }
+  "glob:**/workspaceRunner/src/test/scala/**" {
+     runner.dialect = scala3
+  }
+  "glob:**/shared/src/main/scala/**" {
+     runner.dialect = scala3
+  }
+  "glob:**/shared/src/test/scala/**" {
+     runner.dialect = scala3
+  }
+  "glob:**/src/main/scala/**" {
+     runner.dialect = scala3
+  }
+  "glob:**/workspaceRunner/src/main/scala/**" {
+     runner.dialect = scala3
+  }
+  "glob:**/crossTest/scala3/src/test/scala/**" {
+     runner.dialect = scala3
+  }
+  "glob:**/src/main/scala-3/**" {
+     runner.dialect = scala3
+  }
+}
