+++ log/codegen-trace.md
@@ -1,0 +1,257 @@
+# Agent Execution Trace
+
+**Query:** Create a simple sbt project containing a hello world example that prints the current date and time.  Use 'sbt compile' and 'sbt run' to test the generated code.  You can assume you have sbt and java already installed.Run the program and show the result. 
+**Status:** Failed(Maximum step limit reached)
+**Tools Available:** explore_files, read_file, write_file, modify_file, search_files, execute_command
+
+## Conversation Flow
+
+### Step 1: System Message
+
+```
+You are a helpful assistant with access to tools. 
+Follow these steps:
+1. Analyze the user's question and determine which tools you need to use
+2. Use the necessary tools to find the information needed
+3. When you have enough information, provide a helpful final answer
+4. Think step by step and be thorough
+```
+
+### Step 2: User Message
+
+Create a simple sbt project containing a hello world example that prints the current date and time.  Use 'sbt compile' and 'sbt run' to test the generated code.  You can assume you have sbt and java already installed.Run the program and show the result. 
+
+### Step 3: Assistant Message
+
+To create an sbt project that prints "Hello, World!" along with the current date and time, we'll follow these steps:
+
+1. Create the necessary directory structure for the sbt project.
+2. Write the sbt build file.
+3. Write the Scala code for the "Hello, World!" program.
+4. Compile the project using `sbt compile`.
+5. Run the program using `sbt run`.
+6. Show the output.
+
+Let's begin by setting up the project structure and files.
+
+**Tool Calls:**
+
+Tool: **explore_files**
+
+Arguments:
+```json
+{"path":"./"}
+```
+
+### Step 4: Tool Response
+
+Tool Call ID: `call_2ZrIErRCQKfSAQ6nAxyKUpwT`
+
+Result:
+```json
+{"files":[{"path":"hello-world","type":"directory"},{"path":"my-sbt-project","type":"directory"},{"path":"project","type":"directory"},{"path":"build.sbt","type":"file"},{"path":"src","type":"directory"}]}
+```
+
+### Step 5: Assistant Message
+
+**Tool Calls:**
+
+Tool: **explore_files**
+
+Arguments:
+```json
+{"path":"./my-sbt-project/"}
+```
+
+### Step 6: Tool Response
+
+Tool Call ID: `call_lMcmgG7nPy0nYZRDuO1Vv93Y`
+
+Result:
+```json
+{"files":[{"path":"my-sbt-project/build.sbt","type":"file"},{"path":"my-sbt-project/src","type":"directory"}]}
+```
+
+### Step 7: Assistant Message
+
+**Tool Calls:**
+
+Tool: **explore_files**
+
+Arguments:
+```json
+{"path":"./my-sbt-project/src/"}
+```
+
+### Step 8: Tool Response
+
+Tool Call ID: `call_dtrdhO3irdXcmcIrkW9aVk0i`
+
+Result:
+```json
+{"files":[{"path":"my-sbt-project/src/main","type":"directory"}]}
+```
+
+### Step 9: Assistant Message
+
+**Tool Calls:**
+
+Tool: **explore_files**
+
+Arguments:
+```json
+{"path":"./my-sbt-project/src/main/"}
+```
+
+### Step 10: Tool Response
+
+Tool Call ID: `call_5XkDNA8q3B1oPZT1tvDjQR35`
+
+Result:
+```json
+{"files":[{"path":"my-sbt-project/src/main/scala","type":"directory"}]}
+```
+
+### Step 11: Assistant Message
+
+**Tool Calls:**
+
+Tool: **explore_files**
+
+Arguments:
+```json
+{"path":"./my-sbt-project/src/main/scala/"}
+```
+
+### Step 12: Tool Response
+
+Tool Call ID: `call_7Rl7DZwH64BX4mNQ7qkVjLMM`
+
+Result:
+```json
+{"files":[{"path":"my-sbt-project/src/main/scala/Main.scala","type":"file"}]}
+```
+
+### Step 13: Assistant Message
+
+**Tool Calls:**
+
+Tool: **read_file**
+
+Arguments:
+```json
+{"path":"my-sbt-project/build.sbt"}
+```
+
+### Step 14: Tool Response
+
+Tool Call ID: `call_r4mcDoFFA2qVKiLhfiFhBfT5`
+
+Result:
+```json
+{"content":"name := \"HelloWorld\"\n\nversion := \"0.1\"\n\nscalaVersion := \"2.13.12\""}
+```
+
+### Step 15: Assistant Message
+
+**Tool Calls:**
+
+Tool: **read_file**
+
+Arguments:
+```json
+{"path":"my-sbt-project/src/main/scala/Main.scala"}
+```
+
+### Step 16: Tool Response
+
+Tool Call ID: `call_pCNZeUvCvbkDaan74NfODIyE`
+
+Result:
+```json
+{"content":"import java.time.LocalDateTime\nimport java.time.format.DateTimeFormatter\n\nobject Main extends App {\n  val currentDateTime = LocalDateTime.now()\n  val formatter = DateTimeFormatter.ofPattern(\"yyyy-MM-dd HH:mm:ss\")\n\n  println(\"Hello, World!\")\n  println(\"Current date and time: \" + currentDateTime.format(formatter))\n}"}
+```
+
+### Step 17: Assistant Message
+
+**Tool Calls:**
+
+Tool: **execute_command**
+
+Arguments:
+```json
+{"command":"cd my-sbt-project && sbt compile"}
+```
+
+### Step 18: Tool Response
+
+Tool Call ID: `call_QipyhTOebeKn4o2qVOpV3BDb`
+
+Result:
+```json
+{ "isError": true, "message": "InvalidArguments(List(Exception executing command: TIMEOUT: Command execution timed out after 30000ms))" }
+```
+
+### Step 19: Assistant Message
+
+**Tool Calls:**
+
+Tool: **execute_command**
+
+Arguments:
+```json
+{"command":"cd my-sbt-project && sbt compile"}
+```
+
+### Step 20: Tool Response
+
+Tool Call ID: `call_Clnsg9KbKc1WzkuBamj9rkyj`
+
+Result:
+```json
+{"exit_code":0,"stdout":"[info] welcome to sbt 1.10.11 (Eclipse Adoptium Java 21.0.7)\n[info] loading project definition from /workspace/my-sbt-project/project\n[info] loading settings for project my-sbt-project from build.sbt...\n[info] set current project to HelloWorld (in build file:/workspace/my-sbt-project/)\n[info] Executing in batch mode. For better performance use sbt's shell\n[info] compiling 1 Scala source to /workspace/my-sbt-project/target/scala-2.13/classes ...\n[info] done compiling\n[success] Total time: 6 s, completed May 20, 2025, 7:58:27 AM\n","stderr":""}
+```
+
+### Step 21: Assistant Message
+
+**Tool Calls:**
+
+Tool: **execute_command**
+
+Arguments:
+```json
+{"command":"cd my-sbt-project && sbt run"}
+```
+
+### Step 22: Tool Response
+
+Tool Call ID: `call_zeTGAXV9eqrpyplTqdIlZJ4r`
+
+Result:
+```json
+{"exit_code":0,"stdout":"[info] welcome to sbt 1.10.11 (Eclipse Adoptium Java 21.0.7)\n[info] loading project definition from /workspace/my-sbt-project/project\n[info] loading settings for project my-sbt-project from build.sbt...\n[info] set current project to HelloWorld (in build file:/workspace/my-sbt-project/)\n[info] running Main \nHello, World!\nCurrent date and time: 2025-05-20 07:58:31\n[success] Total time: 0 s, completed May 20, 2025, 7:58:31 AM\n","stderr":""}
+```
+
+## Execution Logs
+
+1. **Assistant:** tools: 1 tool calls requested (explore_files)
+2. **Tools:** executing 1 tools (explore_files)
+3. **Assistant:** tools: 1 tool calls requested (explore_files)
+4. **Tools:** executing 1 tools (explore_files)
+5. **Assistant:** tools: 1 tool calls requested (explore_files)
+6. **Tools:** executing 1 tools (explore_files)
+7. **Assistant:** tools: 1 tool calls requested (explore_files)
+8. **Tools:** executing 1 tools (explore_files)
+9. **Assistant:** tools: 1 tool calls requested (explore_files)
+10. **Tools:** executing 1 tools (explore_files)
+11. **Assistant:** tools: 1 tool calls requested (read_file)
+12. **Tools:** executing 1 tools (read_file)
+13. **Assistant:** tools: 1 tool calls requested (read_file)
+14. **Tools:** executing 1 tools (read_file)
+15. **Assistant:** tools: 1 tool calls requested (execute_command)
+16. **Tools:** executing 1 tools (execute_command)
+17. **Assistant:** tools: 1 tool calls requested (execute_command)
+18. **Tools:** executing 1 tools (execute_command)
+19. **Assistant:** tools: 1 tool calls requested (execute_command)
+20. **Tools:** executing 1 tools (execute_command)
+21. **System:** Step limit reached
