+++ samples/src/main/scala/org/llm4s/samples/toolapi/WorkspaceToolExample.scala
@@ -55,7 +55,7 @@ object WorkspaceToolExample {
         workspace.writeFile("/workspace/another_file.txt", "This is another test file\nWith different content")
 
         // Create the workspace tools
-        val workspaceTools = createWorkspaceTools(workspace)
+        val workspaceTools = WorkspaceTools.createDefaultWorkspaceTools(workspace)
         val toolRegistry = new ToolRegistry(workspaceTools)
 
         // Create test prompt for the LLM
@@ -99,162 +99,6 @@ object WorkspaceToolExample {
     }
   }
 
-  /**
-   * Creates a list of tools that wrap workspace functionality
-   */
-  private def createWorkspaceTools(workspace: ContainerisedWorkspace): Seq[ToolFunction[_, _]] = {
-    // 1. Tool for exploring files
-    val exploreSchema = Schema
-      .`object`[Map[String, Any]]("Explore files parameters")
-      .withProperty(
-        Schema.property(
-          "path",
-          Schema.string("Path to explore in the workspace")
-        )
-      )
-      .withProperty(
-        Schema.property(
-          "recursive",
-          Schema.boolean("Whether to explore recursively")
-        )
-      )
-
-    def exploreHandler(params: SafeParameterExtractor): Either[String, ExploreResult] =
-      for {
-        path <- params.getString("path")
-        recursiveValue = params.getBoolean("recursive").toOption.getOrElse(false)
-      } yield {
-        val result = workspace.exploreFiles(path, recursive = Some(recursiveValue))
-        val files = result.files.filter(!_.isDirectory).map(_.path)
-        val dirs = result.files.filter(_.isDirectory).map(_.path)
-        ExploreResult(files, dirs)
-      }
-
-    val exploreTool = ToolBuilder[Map[String, Any], ExploreResult](
-      "explore_files",
-      "Explore files and directories in the workspace",
-      exploreSchema
-    ).withHandler(exploreHandler).build()
-
-    // 2. Tool for reading files
-    val readSchema = Schema
-      .`object`[Map[String, Any]]("Read file parameters")
-      .withProperty(
-        Schema.property(
-          "path",
-          Schema.string("Path to the file to read")
-        )
-      )
-      .withProperty(
-        Schema.property(
-          "startLine",
-          Schema.integer("Starting line (1-indexed)")
-        )
-      )
-      .withProperty(
-        Schema.property(
-          "endLine",
-          Schema.integer("Ending line (1-indexed)")
-        )
-      )
-
-    def readHandler(params: SafeParameterExtractor): Either[String, ReadResult] =
-      for {
-        path <- params.getString("path")
-        startLineOpt = params.getInt("startLine").toOption
-        endLineOpt = params.getInt("endLine").toOption
-      } yield {
-        val result = workspace.readFile(path, startLineOpt, endLineOpt)
-        ReadResult(result.content, result.totalLines)
-      }
-
-    val readTool = ToolBuilder[Map[String, Any], ReadResult](
-      "read_file",
-      "Read content from a file in the workspace",
-      readSchema
-    ).withHandler(readHandler).build()
-
-    // 3. Tool for searching in files
-    val searchSchema = Schema
-      .`object`[Map[String, Any]]("Search files parameters")
-      .withProperty(
-        Schema.property(
-          "query",
-          Schema.string("Text to search for")
-        )
-      )
-      .withProperty(
-        Schema.property(
-          "paths",
-          Schema.array("List of paths to search in", Schema.string("Path to search in"))
-        )
-      )
-      .withProperty(
-        Schema.property(
-          "searchType",
-          Schema.string("Type of search (regex or literal)")
-            .withEnum(Seq("regex", "literal"))
-        )
-      )
-      .withProperty(
-        Schema.property(
-          "recursive",
-          Schema.boolean("Whether to search recursively")
-        )
-      )
-
-    def searchHandler(params: SafeParameterExtractor): Either[String, SearchResult] =
-      for {
-        query <- params.getString("query")
-        pathsArr <- params.getArray("paths")
-        pathsList = pathsArr.arr.map(_.str).toList
-        searchType <- params.getString("searchType")
-        recursiveValue = params.getBoolean("recursive").toOption.getOrElse(true)
-      } yield {
-        val result = workspace.searchFiles(pathsList, query, searchType, recursive = Some(recursiveValue))
-        val matchStrings = result.matches.map(m => s"${m.path}:${m.line}: ${m.matchText}")
-        SearchResult(matchStrings, result.totalMatches)
-      }
-
-    val searchTool = ToolBuilder[Map[String, Any], SearchResult](
-      "search_files",
-      "Search for content in files in the workspace",
-      searchSchema
-    ).withHandler(searchHandler).build()
-
-    // 4. Tool for executing commands
-    val executeSchema = Schema
-      .`object`[Map[String, Any]]("Execute command parameters")
-      .withProperty(
-        Schema.property(
-          "command",
-          Schema.string("Command to execute")
-        )
-      )
-      .withProperty(
-        Schema.property(
-          "workingDirectory",
-          Schema.string("Working directory")
-        )
-      )
-
-    def executeHandler(params: SafeParameterExtractor): Either[String, ExecuteResult] =
-      for {
-        command <- params.getString("command")
-        workingDir = params.getString("workingDirectory").toOption
-      } yield {
-        val result = workspace.executeCommand(command, workingDir)
-        ExecuteResult(result.stdout + "\n" + result.stderr, result.exitCode)
-      }
-
-    val executeTool = ToolBuilder[Map[String, Any], ExecuteResult](
-      "execute_command",
-      "Execute a command in the workspace",
-      executeSchema
-    ).withHandler(executeHandler).build()
-
-    Seq(exploreTool, readTool, searchTool, executeTool)
-  }
 
   /**
    * Test an LLM with the workspace tools
