+++ src/main/scala/org/llm4s/llmconnect/provider/AnthropicClient.scala
@@ -107,14 +107,12 @@ curl https://api.anthropic.com/v1/messages \
     // Track if we've seen a system message
     var hasSystemMessage = false
 
-
     println("\n\n********************************************************************************************")
     println("********************************************************************************************")
     println("Got conversation: " + conversation.messages.size + " messages to encode")
     // Process messages in order
     conversation.messages.foreach {
       case SystemMessage(content) =>
-
         println("\n\nXXXX Adding system message: " + content)
         paramsBuilder.system(content)
         hasSystemMessage = true
